Remove date range requirement for STA orders
STA orders to not necessarily need a date range. Actually the default should be to not submit it. When requesting without date range, the remote server only returns those orders which have not yet been retrieved. A date range is only needed if already received statements should be fetched again.