Adapt to BREAKING change in faraday v0.16.0
The faraday gem has decided to differentiate between request body
and response body, thereby dropping the use of the `:body` key in `env`

https://github.com/lostisland/faraday/commit/6e3c40f159ef514064d05faabf6c6f261662970b

This is in perfect alignment with semantic versioning as breaking
changes in minor increments are legal before v1.0.0.

As faraday `>= 0.16.0` is only available for Ruby `>= 2.3`
the change broke only the builds for newer Ruby versions.