Fix order data encryption for OpenSSL 3
With OpenSSL 3, `cipher.reset` requires an explicit `iv` value to generate
identical values. Using empty initialization vector yields values identical
to the OpenSSL 1 values.

In addition, make sure that the cipher is set to `encrypt` right after
initialization which is required before calling `random_key` according to
the docs. The call to `encrypt` doesn't need to be repeated as `reset`
keeps the previous mode.